requirements.txt
accelerate==0.21.0
bitsandbytes==0.43.1
decord==0.6.0
einops==0.8.0
huggingface_hub==0.24.6
icecream==2.1.3
iopath==0.1.10
matplotlib==3.8.4
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.57.4
opencv_contrib_python==4.5.4.60
opencv_python==4.5.4.60
opencv_python_headless==4.5.4.60
pandas==2.2.3
peft==0.13.2
pillow==10.3.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
pycocoevalcap
pycocotools
PyYAML==6.0.1
Requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
scikit-image
sentencepiece==0.2.0
tiktoken==0.8.0
timm==1.0.12
tqdm==4.66.5
transformers==4.37.2
visual_genome==1.1.1
wandb
webdataset